Fix caret focusing issues while navigating help
There was a regression (caused by a change the underlying code in
Sublime itself) that "broke" the code that focuses the cursor on anchors
while navigating through them.

This is fixed by properly using a text command to change the selection
state instead of using the old hack that forces a window redraw, which
on the whole is more robust.