Fix post processing to be less dumb
While adding in support for comments (and removing them from the buffer)
I noticed that I was doing something silly and making several methods
toggle the read-only state of the buffer.

There may have been historic reasons for this (perhaps they used to not
be module specific, for example), but even so this seems silly at this
point.

This is now fixed and things look a little cleaner.