Ensure that links are reclassified on load
Now that links can be marked as both active and broken, it's important
to ensure that when hyperhelp loads, it re-checks all links to ensure
that the highlight is correct.

In particular, it's important to note that the regions used are
persistent in the session file, but the state of the help index on disk
may not be the same as it was before.