Include new helper command to rewrite anchors
This modifies how we post process anchors so that instead of doing it
via selection modifications we instead use simple region replacement.
In order to make this possible, the syntax also has been enhanced so 
that the punctuation on hidden anchors is more explicit than the ones
for "regular" anchors.

The command does what the old one did, but in fewer lines of code.

One side effect of this is that the structure of the setting used to
track the anchors in the file has been changed from being a list of
lists to being a dictionary that indexes regions, which is faster and
cleaner.