Finalize initial implementation of new anchors
This finishes up the new implementation of anchors by including the
ability to have the text of an anchor be different than the topic ID
that it's associated with.

With this in place, we can now create anchors that link to a unique ID
but seem to have alternate text.