Perform some simple cleanup of the post processor
This moves the internal commands for doing post processing into their
own file and includes an extra safety mechanism of both including the
text `internal` in their command name as well as requiring that an extra
setting be set so the command knows that it should do something.

This should hopefully keep our help files safe from both curious users
as well as ourselves if we accidentally do something dumb.