Replace the link post processor with a command
The old command for post processing links has been replaced with a new
internal command for doing the same thing.

In preparation for an upcoming feature, this new command creates a new
setting that will be used to associate link positions in the buffer with
the topics and packages that the links should ultimately navigate to.