Implement full link rewriting using new parser
This implements into the post processing for links the new parser for
link bodies, which causes it to re-write all of the links that have a
package, topic, or both specified uniquely, setting up the appropriate
underlying setting so that the link following still works.

Note that there are still issues with event handling, so the hover
popup will incorrectly tell you link destinations, and active/broken
links don't get reassessed on startup currently, although they should.