Rename the internal commands that post process
For clarity, the names of the two existing commands whose job it is to
post process things have been renamed to include `Process` in their name
instead of `Capture`.

Visually, this makes a bit more sense and will be in line with some new
upcoming internal commands that don't do this.