Make a common post processing enable check
This shifts the code that knows if a post processing command can execute
from `is_enabled` to a module level helper function so that all commands
for post processing can leverage a single call point and we can change
the criteria for everything at once if needed.

Probably the better way to go would be a simple common base class that
directly implements `is_enabled()` with the appropriate logic, but for
now we're going this way while we determine if maybe there aren't other
possible internal commands that would be controlled a slightly different
way as well as this one, in which case this may be a better solution.