Factor link flagging code into its own command
This pulls the part of the internal command for processing links that
determines if a link points to a valid destination or not out of the
existing command and into a command of it's own.

This new command only concerns itself with applying the appropriate
regions to links based on looking up their topics, so that this can be
done at any point.

The post processing command for links invokes this command as part of
its own operation, so externally this works the same as it did
previously.