Re-enable the event handler for flagging links
Prior to reworking links, there was a plugin_loaded that wuold do a
post process of links to mark some of them as active and others as
broken.

This was temporarily removed during the link reworking because the
post processing was switching to being destructive in order to allow for
link text being different from the topic.

Now that the flagging code is separate from the rest of the post
processing, this can be included again.