Change the _hh_links setting structure
Where previously this was a dictionary where the index was a region
start position (encoded as a string) and the value was the topic data
dictionary, the setting now uses a region index instead of position.

The general idea is that although the position of the region may change,
the location of the region in the buffer does not.

This makes link navigation more robust in cases where the text of the
loaded help file gets modified after links are post processed.