Implement some new hover handling
This modifies the on_hover handler so that it looks up topics using the
new topic mechanism.

Since links may now link to other packages, the contents of the hover
popup has been modified somewhat from it's initial version to see if we
can get it looking a bit nicer and providing more information.

In particular, we want to display as much important information as is
possible, so all links display information to let you know what package,
file and topic are being navigated to.

In order to not overload anyone, this information display elides any
parts that are not necessary, so package is not displayed if it's the
current package and file is not displayed if it's the current file.

In the case of a broken link, package and topic are always displayed so
as to provide more trouble shooting information.