Change the hh_links setting from array to dict
After working with the new system a bit, I don't like how this is using
a dictionary where the keys are str()'d ints when it could just be an
array since in most cases we need to look things up by region and not
their index anyway.

So now the hh_links setting is an array where the elements are directly
the topic data and they share the same indexes with regions in the
region list, so if you know what the region is, you know what the index
is, and vice versa.