Update _get_link_topic for the new setting type
This makes some slight modifications to the helper function for getting
topic data based on a link region or link index so that it accesses the
data through the new setting.

Since all other accesses go through this (following links, hover popup,
link flagging) this fixes all link actions in one shot which being a
slight optimization in cases where a link index is provided instead of a
region.