Optimize link flagging
Flagging of links as active or broken is a candidate for using the
lookup API to use an index instead of a region, since it steps through
all regions in order anyway.

This was actually the original intention but I forgot to come back and
tweak it before the original series of commits.