Mark links to missing package files as broken
This simple change enhances link flagging so that when it is flagging
links that will open a package file, a check is done to see if that
package file exists or not.

Links that will open a package file that don't exist as a resource will
be flagged as broken.