Clean up the main context
As a first step to cleaning up the help syntax, this change expands the
main context that is responsible for matching headers in files out to
sub contexts for extra clarity when reading the syntax.

This also modifies the regex for the title of the help page so that it
doesn't include trailing whitespace and removes some variables that
probably don't actually need to be variables.