Include new utilities for classifying topics
Since this comes up in many places both here in the core as well as in
the authoring tools, the API for the core has been extended to include
some extra routines that can classify topics.

Generally topics are one of "normal", "file" or "url", that is they
either represent a specific topic, they open a package file, or they
open an external URL.

The added API functions allow for both checking which of the three
types the link is as well as validating (or files or URL's) that the
underlying data appears to be valid.

For a package file, this means verifying that such a file currently
exists, while for a URL it is a verification that the URL appears to be
a URL.