Use the new API's for some internal tests
Now that there is an "official" API for checking the state of a topic
to see what it is and if it's valid, this is now used in the link flag
code and in the hover popup, which simplifies them a bit.

In both cases we take advantage of knowing that the function for
validating that a package file exists will return None for things that
are not links to package files by using `is` to verify that we only
handle a `False` value as False.