Modify the index command to allow initial text
This is a simple modification to the exiting command that opens the help
index for either a specific package or one the user species so that the
text initially entered into the quick panel can be pre-entered for the
user.

SnAPI uses this to display partial patches on topics when the index
contains more than one matching, relevant topic.