Rename the core syntaxes
This alters the names of the files that carry the core syntaxes so that
we can clear up a settings file in the name of the package for future
features.

This is primarily a cosmetic change, although a bit of code that
programatically assigns syntaxes needed to be changed as well.