Include a setting for storing user bookmarks
This includes a new setting called bookmarks for storing a list of
topics that the user has decided to bookmark. Each bookmark contains
information on the topic and package being bookmarked, a name for
display in menus and the command palette as well as optional cursor and
viewport information.

Included in this commit is a small set of potential API functions for
working with bookmarks while we flesh out what the API should look like.