Include the ability to modify a setting
For an upcoming change, and for orthogonality with the method for
getting the value of a setting, this implements the ability to add or
alter a setting in the HyperHelp settings object.

Along with setting the value, this also allows for causing the modified
setting to be persisted back to disk.