Include input handlers for bookmark type and name
For an upcoming command, this implements an initial revision of some
command input handlers that allow the user to select what style of
bookmark they would like to create as well as the name to use for that
bookmark.

These should probably be reworked (along with the command that is coming
up) to provide more control over the exact bookmark to create, but for
the time being this is good enough while we evaluate the current
implementation.