Include an initial command for bookmark creation
This implements a command that's context sensitive to the state of the
current view (requiring a help view) and gathers information for the
creation of a new bookmark.

The command defers the actual creation of the bookmark to a command that
has not yet been committed.

Whoops

Missed this.

double whoops