Implement a command to create a bookmark
This uses arguments provided (which it does not validate for accuracy)
to use the existing API method for adding a new bookmark to the list of
bookmarks for the current user.

This command is invoked by the previously committed command; unlike that
command this command has no context sensitivity and just does what you
tell it.