Include a context command for bookmarking
This includes a context sensitive command that is only available in a
help view and which allows you to bookmark either the current file or
the topic that a link points to, depending on the location of the
context event.

This may need a bit of tweaking; it currently fiddles with the selection
if you choose to bookmark a link in order to make the existing command
do what is needed.

In all likelihood that command also needs to change, but this is enough
for the time being while we evaluate our bookmark solution.