Modify anchors, links and keybinds to allow spaces
Where previously these constructs required the use of a non-breaking
(unicode) space character in order to have visible whitespace, the
match rules have now been extended.

Instead, the only rule in place for whitespace is now that it can't
immediately follow the token that opens the construct up.

This should hopefully be way more intuitive than it used to be while
at the same time not causing an ambiguity in most casts.