Remove all redundant NBSP space handling in index
Where previously the index source file used spaces to represent white
space in topics and that was changed to a non-breaking space character
on index load so that it would match the contents of the actual help
buffer, this is no longer needed.

Now the help buffer can contain regular white pace, which matches what
the index already contains, so there's no need to fiddle with it.

In the place of that we ensure that we coalesce sequence of two or more
whitespace characters into a single one, since it's now possible for
anchors and links to straddle line boundaries, and newlines are white
space.