Ensure that parsed links don't have space issues
Since it's now possible to have internal whitespace in anchors and in
links, it seems like a good idea to modify the code that parses the body
of a link and an anchor to strip away the white space characters that
might be leading or trailing on any of the parts.

In particular this is important because it's possible to generate a link
or an anchor that, when rewritten, is no longer syntactically identified
as the correct type of item due to leading space.