Allow for empty headers
The rule for matching a heading anchor ran afoul of a particularly
nasty situation in which a single `#` on a line makes the syntax go
into header mode and not leave if the remainder of the line doesn't
include any topic data.

The result is that a stray `#` on a line by itself renders the remainder
of the help file un-highlighted.

To remedy this a simple escape match has been included in the
header-anchors context to allow the context to escape if a header is not
found.

This is probably not the correct way to do this, but since there are
plans to circle back and fix the syntax, this is a good enough stop gap
measure for now considering the current quality of the syntax as a
whole.