Fix the settings object on initial bootstrap
The `hh_setting` API method for getting a setting from the hyperhelp
config file had a weird race condition type thing wherein due to the
fact that it was relying on attributes being initialized in
`plugin_loaded`, sometimes that would get called and sometimes it would
not, depending on load order.

The appropriate attributes are now lazily initialized on first call
instead.