Convert the package into a full blown dependency
This completes the last major step in turning the code into a proper
dependency instead of just a test package by setting up the appropriate
structure in the package so that the PackageControl dependency installer
knows how to handle it.

This is primarily just moving files into an `all/hyperhelpcore` folder,
although along with this we also renamed the internal help index to 
have an appropriate ignored suffix like other resources to make sure
that hyperhelp doesn't try to load it, since it results in a duplicate
index that is non-cannonical.