Set the initial dependency version information
This sets our initial version tuple to 0.0.1, which is as far as I know
the smallest possible version, or at least the smallest semver that 
makes any sense.

From this point forward, changes to anything that we want anyone to
see need to have the version tuple bumped and a new release created.