Add package filtering to help index scans
The internal `_scan_help_packages()` function is responsible for finding
all help index files and loading those that have not previously been
seen.

This changes introduces an (optional) filter for this function to allow
it to only find and load help indexes contained with specific subsets of
packages.

This allows for more granular control of help index detection and allows
us to demand load all of the help indexes from specific packages if
desired.