hyperhelp_allow_unpacked should allow bootstraps
As originally written, the `hyperhelp_allow_unpacked` setting would turn
the error about there being an unpacked version of the bootstrapped
package into just a warning, but it would still block a bootstrap
check.

The spirit of this setting is to turn this from a blocking error to a
logged warning but still proceed. So the logic has been altered so that
when the setting is enabled, the check to see if we should bootstrap
proceeds instead of being blocked.