Watch the ignored_package setting and update index
This code in the event handler of the bootstrapped system package for
HyperHelp uses the newly added functionality in the core for loading and
unloading help indexes from physical packages to keep the help indexes
up to date as packages are installed and removed.

To do this, the list of ignored packages is monitored for changes, and
any packages added to the list have their help indexes (if any) unloaded
while packages removed from the list have their packages (if any) added.

This also bumps the internal version number up since the package
contents have changed.