Enhance context handling with expanded universe
In order to be more generally useful, as well as to provide a layer of
insulation should we decide to change up the syntaxes, we have now
implemented a set of contexts to be used in place of the single one
that previously existed.

These allow us to check things like whether a file is help source or not
without having to verify with selectors, and also includes an enhanced
set such as being able to determine if help is open at all and, if so,
what package it's currently displaying.