Fix the history quick panel index setting
This tweaks up the quick panel that is used to show the history so that
instead of just showing you which entry is the one you're currently
viewing by tagging it, we also now set that item to be the selected
index when the panel opens.

This gives a little bit more context regarding where in the list of
history you are, allowing you to scroll up or down as appropriate.

One downside of this is that if there are many history entries and the
current one is closer to the bottom of the list, Sublime opens the panel
with the selected item at the bottom and not the middle, but as far as
I can see that's not really something we can deal with here.