Fix syntax embedding in code blocks
This commit features two fixes for the syntax in relation to embedding
the syntax for other languages for use in code blocks.

First, since we embed Python for code blocks, and Python methods and
classes appear in the symbol list, we have added an extra symbol list
setting that excludes python symbols when they appear embedded in help
files so that we don't pollute the symbol list with them (since that is
where heading go).

Additionally, the code for redacting comments from loaded help files was
too loose in it's scope definition, as it would also clobber comments in
all of the embedded syntaxes (JSON, Python, etc). We now ensure that
we don't do that by only finding and removing help comments instead.