Fix the topic used for help on help
This fixes the topic used for the help on help menu command so that
it actually provides help.

I suspect that this was done in preparation for swapping the contents of
index and help_on_help (and renaming the latter) so that the entry point
to the help is the help on help, but I'm not sure.