Complete revamp of the internal help
This is a large scale revamp of the internal help, which had not been
officially touched in a long time while heavy development was going on.

Where previously the only changes made were to keep existing help
following the evolving syntax of the package and it's commands, this
change encompasses bringing everything as up to date as possible with
the current state of the package, including a description of how it
works internally and how to integrate it with packages.

All that is currently left undocumented (but planned) is help for the
API, which is not being done currently because until the HyperHelpAuthor
changes are finalized, it may be subject to change. And also to some
small degree, I'm kind of tired of writing help at the moment. But
mostly that first one.