Include developer mode commands
This includes extra commands into the HyperHelp bootstrap package that
are meant for developers only. They disable themselves for anyone that's
not a developer due to the potential for mischief.

Currently the only such command is a command to force a bootstrap to
occur right now, which is handy for development purposes.