Include a command for displaying version info
In preparation for letting people other than those following the live
streams and those who found the repository and might be using it already
that HyperHelp and the related packages are a thing, this version adds
in an about command.

This currently displays the version number of the dependency and the
version number of the bootstrapped package, which will hopefully be
useful in case anyone has issues from this point forward.

THis is also flagged as a version version as well, so that anyone that
already has the code will get the latest version.