Namespace our custom user preferences differently
Previously our custom settings were prefixed with `hyperhelp_`, but they
are now prefixed with `hyperhelp.` instead, which seems to be a more
accepted standard for this (and PackageDev highlights them differently
as well).

In addition to this, a new hint file has been added so that PackageDev
can autocomplete the settings for the user and also provide information
on the default values of each setting and what it does.