Use show_overlay to execute bookmark commands
There are known issues in Sublime in which invoking a command that uses
input handlers via a key binding may cause the internals to go a little
haywire and execute the wrong command.

Anecdotally, it seems like this happens less frequently if you use the
`show_overlay` command to open the command palette and tell it to
execute the command rather than invoking the command directly.

Additionally, it also seems to be the more "correct" way to do this
(it's what Sublime Merge does, for example), so it seems like a best
practice sort of change in general.