Include some PackageDev integrations
The PackageDev package provides special syntax highlighting for settings
keys that are namespaced based on their package names, so our internal
settings for the dependency section have been modified to use that
scheme to take advantage of this. In addition we also now ship a hints
file so that PackageDev can provide autocompletion of our settings as
well as hover popups that describe them and their default values.

Note that officially only one of the internal dependency settings is
documented, although all of them have been added to the hints file. This
is on purpose because generally speaking the not-officially-documented
settings aren't ones that people should be using except in very specific
circumstances, and those settings are explained when they're needed in
popups.

The changelog was also modified to bring up to date the changes for this
version.