Allow the ability to display changelog on update
To finalize the release for today, this adds in the ability of the help
package to automatically display a help topic upon package reload after
the bootstrap has either installed or updated the package.

On an initial install, the help on help topic is displayed while for
updates the changelog is displayed instead, focusing on the most recent
set of changes.

In both cases, the new setting `show_changelog` controls whether this is
displayed or not, with the default being `True`. Turing this off in your
settings will stop HyperHelp from displaying the help topic.