Augment the hyperhelp_focus command logic
Previously this command sported only an `at_center` argument which would
determine if the focused areas would be centered in the viewport or just
made visible.

The command now also takes an `at_top` argument which, if provided takes
precedence and focuses the region to the top of the viewport in the
window instead.