Include a new setting, focus_links_at_top
This setting controls what happens to the focus when you follow a link.
If this setting is set to `true` (which is the default), following a
link will put the link at the top of the window; otherwise it will be
centered, which is how this used to work.

Arguably when you follow a link you want to see as much text associated
with it as possible, so this should probably have been the default for
this long ago; the centered option is only for backwards compatibility
with prior versions.